Fixed the DTO problem on Barber Coupon API.
Previously the web app UI was not able to show the barber coupons.
This is because the DTOs were missing on the coupon API project.
Updated the coupon API project with DTO's.
Now web app UI is able to get the data and also show the data as a list in the UI.